+++ app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/addiction/AddIncomeHistoryFragment.kt
@@ -1,4 +1,4 @@
-package com.romandevyatov.bestfinance.ui.fragments
+package com.romandevyatov.bestfinance.ui.fragments.addiction
 
 import android.app.DatePickerDialog
 import android.content.Context
@@ -27,7 +27,6 @@ import dagger.hilt.android.AndroidEntryPoint
 import java.time.OffsetDateTime
 import java.time.format.DateTimeFormatter
 import java.util.*
-
 
 @AndroidEntryPoint
 class AddIncomeHistoryFragment : Fragment() {
