+++ app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/addiction/AddIncomeSubGroupFragment.kt
@@ -1,4 +1,4 @@
-package com.romandevyatov.bestfinance.ui.fragments
+package com.romandevyatov.bestfinance.ui.fragments.addiction
 
 import android.graphics.Color
 import android.os.Bundle
@@ -17,7 +17,6 @@ import com.romandevyatov.bestfinance.db.entities.IncomeSubGroup
 import com.romandevyatov.bestfinance.viewmodels.foreachmodel.IncomeGroupViewModel
 import com.romandevyatov.bestfinance.viewmodels.foreachmodel.IncomeSubGroupViewModel
 import dagger.hilt.android.AndroidEntryPoint
-
 
 @AndroidEntryPoint
 class AddIncomeSubGroupFragment : Fragment() {
