+++ app/src/main/java/com/romandevyatov/bestfinance/ui/validators/ValidationCodeConstants.kt
@@ -1,4 +1,4 @@
-package com.romandevyatov.bestfinance.ui.validator
+package com.romandevyatov.bestfinance.ui.validators
 
 object ValidationCodeConstants {
