+++ app/src/main/java/com/romandevyatov/bestfinance/ui/validators/base/BaseValidator.kt
@@ -1,4 +1,4 @@
-package com.romandevyatov.bestfinance.ui.validator.base
+package com.romandevyatov.bestfinance.ui.validators.base
 
 import com.romandevyatov.bestfinance.R
