+++ app/src/main/res/layouts/fragments/layout/fragment_add_expense_group.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/addExpenseGroupNameFragment"
-    tools:context=".ui.fragments.AddExpenseGroupNameFragment" >
+    tools:context=".ui.fragments.addiction.AddExpenseGroupNameFragment" >
 
     <EditText
         android:id="@+id/newExpenseGroupName"
