+++ app/src/main/res/layouts/fragments/layout/fragment_add_expense_sub_group.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/addExpenseSubGroupFragment"
-    tools:context=".ui.fragments.AddExpenseSubGroupFragment">
+    tools:context=".ui.fragments.addiction.AddExpenseSubGroupFragment">
 
     <EditText
         android:id="@+id/newExpenseSubGroupName"
