+++ app/src/main/res/layouts/fragments/layout/fragment_add_income_group.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/addIncomeGroupFragment"
-    tools:context=".ui.fragments.AddIncomeGroupFragment">
+    tools:context=".ui.fragments.addiction.AddIncomeGroupFragment">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/newIncomeGroupNameInputLayout"
