+++ app/src/main/res/layouts/fragments/layout/fragment_add_income_history.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/addIncomeHistoryFragment"
-    tools:context=".ui.fragments.AddIncomeHistoryFragment" >
+    tools:context=".ui.fragments.addiction.AddIncomeHistoryFragment" >
 
     <Spinner
         android:id="@+id/incomeGroupSpinner"
