+++ app/src/main/res/layouts/fragments/layout/fragment_add_income_sub_group.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/addIncomeSubGroupFragment"
-    tools:context=".ui.fragments.AddIncomeSubGroupFragment" >
+    tools:context=".ui.fragments.addiction.AddIncomeSubGroupFragment" >
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/newIncomeSubGroupNameContainer"
