+++ app/src/main/res/layouts/fragments/layout/fragment_analyze.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:id="@+id/addWalletFragment"
-    tools:context=".ui.fragments.AddWalletFragment">
+    tools:context=".ui.fragments.addiction.AddWalletFragment">
 
     <TextView
         android:id="@+id/analyzeGroupTextView"
