+++ app/src/main/res/layouts/fragments/layout/fragment_transfer.xml
@@ -7,7 +7,7 @@
     android:layout_marginStart="10dp"
     android:layout_marginEnd="10dp"
     android:id="@+id/addTransferFragment"
-    tools:context=".ui.fragments.AddTransferFragment">
+    tools:context=".ui.fragments.addiction.AddTransferFragment">
 
 
     <Spinner
