+++ app/src/main/res/navigation/nav_graph.xml
@@ -39,7 +39,7 @@
 
     <fragment
         android:id="@+id/navigation_add_income"
-        android:name="com.romandevyatov.bestfinance.ui.fragments.AddIncomeHistoryFragment"
+        android:name="com.romandevyatov.bestfinance.ui.fragments.addiction.AddIncomeHistoryFragment"
         android:label="AddIncome"
         tools:layout="@layout/fragment_add_income_history" >
         <action
@@ -75,7 +75,7 @@
 
     <fragment
         android:id="@+id/navigation_add_expense"
-        android:name="com.romandevyatov.bestfinance.ui.fragments.AddExpenseHistoryFragment"
+        android:name="com.romandevyatov.bestfinance.ui.fragments.addiction.AddExpenseHistoryFragment"
         android:label="fragment_add_expense_history"
         tools:layout="@layout/fragment_add_expense_history" >
         <action
@@ -114,7 +114,7 @@
         tools:layout="@layout/fragment_history" />
     <fragment
         android:id="@+id/navigation_add_new_income_group"
-        android:name="com.romandevyatov.bestfinance.ui.fragments.AddIncomeGroupFragment"
+        android:name="com.romandevyatov.bestfinance.ui.fragments.addiction.AddIncomeGroupFragment"
         android:label="fragment_add_new_income_group"
         tools:layout="@layout/fragment_add_income_group" >
 
@@ -124,7 +124,7 @@
     </fragment>
     <fragment
         android:id="@+id/navigation_add_new_income_sub_group"
-        android:name="com.romandevyatov.bestfinance.ui.fragments.AddIncomeSubGroupFragment"
+        android:name="com.romandevyatov.bestfinance.ui.fragments.addiction.AddIncomeSubGroupFragment"
         android:label="fragment_add_new_income_sub_group"
         tools:layout="@layout/fragment_add_income_sub_group" >
 
@@ -140,7 +140,7 @@
     </fragment>
     <fragment
         android:id="@+id/navigation_add_new_expense_group"
-        android:name="com.romandevyatov.bestfinance.ui.fragments.AddExpenseGroupNameFragment"
+        android:name="com.romandevyatov.bestfinance.ui.fragments.addiction.AddExpenseGroupNameFragment"
         android:label="fragment_add_new_expense_group"
         tools:layout="@layout/fragment_add_expense_group" >
 
@@ -151,7 +151,7 @@
 
     <fragment
         android:id="@+id/navigation_add_new_expense_sub_group"
-        android:name="com.romandevyatov.bestfinance.ui.fragments.AddExpenseSubGroupFragment"
+        android:name="com.romandevyatov.bestfinance.ui.fragments.addiction.AddExpenseSubGroupFragment"
         android:label="fragment_add_new_expense_sub_group"
         tools:layout="@layout/fragment_add_expense_sub_group" >
 
@@ -168,7 +168,7 @@
 
     <fragment
         android:id="@+id/navigation_add_wallet"
-        android:name="com.romandevyatov.bestfinance.ui.fragments.AddWalletFragment"
+        android:name="com.romandevyatov.bestfinance.ui.fragments.addiction.AddWalletFragment"
         android:label="Add Wallet"
         tools:layout="@layout/fragment_add_wallet" >
         <action
@@ -189,7 +189,7 @@
     </fragment>
     <fragment
         android:id="@+id/navigation_add_new_transfer"
-        android:name="com.romandevyatov.bestfinance.ui.fragments.AddTransferFragment"
+        android:name="com.romandevyatov.bestfinance.ui.fragments.addiction.AddTransferFragment"
         android:label="fragment_transfer"
         tools:layout="@layout/fragment_transfer" >
         <action
